pyproject.toml
[tool.poetry]
name = "the-wall-api"
version = "0.1.0"
description = "The Wall DRF API"
authors = ["Aleksandar Dimitrov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
drf-spectacular = "^0.27.2"
openapi-spec-validator = "^0.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
